Move _make_binary_structure to utils module
